filter: Support a float time value for trigger
Currently --time-filter option support float time as below.

  $ uftrace -t 0.3us ./a.out

But the time tigger cannot handle a float value
so fix it using existing parse_time().

Before:
  $ uftrace record ./hello
  $ uftrace replay -T main@time=1.79us
  # DURATION    TID     FUNCTION
     1.013 us [19810] | __monstartup();
     0.649 us [19810] | __cxa_atexit();
              [19810] | main() {
    11.739 us [19810] |   puts();
              [19810] |   foo() {
     1.785 us [19810] |     puts();
     2.357 us [19810] |   } /* foo */
    15.040 us [19810] | } /* main */

After:
  $ uftrace replay -T main@time=1.79us
  # DURATION    TID     FUNCTION
     1.013 us [19810] | __monstartup();
     0.649 us [19810] | __cxa_atexit();
              [19810] | main() {
    11.739 us [19810] |   puts();
     2.357 us [19810] |   foo();
    15.040 us [19810] | } /* main */
