+++ uftrace.c
@@ -292,101 +292,6 @@ static void parse_debug_domain(char *arg)
 	free(saved_str);
 }
 
-static int get_digits(uint64_t num)
-{
-	int digits = 0;
-
-	do {
-		num /= 10;
-		digits++;
-	} while (num != 0);
-
-	return digits;
-}
-
-static uint64_t parse_min(uint64_t min, uint64_t decimal, int decimal_places)
-{
-	uint64_t nsec = min * 60 * NSEC_PER_SEC;
-
-	if (decimal) {
-		decimal_places += get_digits(decimal);
-		decimal *= 6;
-
-		/* decide a unit from the number of decimal places */
-		switch (decimal_places) {
-		case 1:
-			nsec += decimal * NSEC_PER_SEC;
-			break;
-		case 2:
-			decimal *= 10;
-		case 3:
-			decimal *= 10;
-			nsec += decimal * NSEC_PER_MSEC;
-			break;
-		default:
-			break;
-		}
-	}
-	return nsec;
-}
-
-static uint64_t parse_time(char *arg, int limited_digits)
-{
-	char *unit, *pos;
-	int i, decimal_places = 0, exp = 0;
-	uint64_t limited, decimal = 0;
-	uint64_t val = strtoull(arg, &unit, 0);
-
-	pos = strchr(arg, '.');
-	if (pos != NULL) {
-		while (*(++pos) == '0')
-			decimal_places++;
-		decimal = strtoull(pos, &unit, 0);
-	}
-
-	limited = 10;
-	for (i = 1; i < limited_digits; i++)
-		limited *= 10;
-	if (val >= limited)
-		pr_err_ns("Limited %d digits (before and after decimal point)\n",
-			  limited_digits);
-	/* ignore more digits than limited digits before decimal point */
-	while (decimal >= limited)
-		decimal /=10;
-
-	/*
-	 * if the unit is omitted, it is regarded as default unit 'ns'.
-	 * so ignore it before decimal point.
-	 */
-	if (unit == NULL || *unit == '\0')
-		return val;
-
-	if (!strcasecmp(unit, "ns") || !strcasecmp(unit, "nsec"))
-		return val;
-	else if (!strcasecmp(unit, "us") || !strcasecmp(unit, "usec"))
-		exp = 3; /* 10^3*/
-	else if (!strcasecmp(unit, "ms") || !strcasecmp(unit, "msec"))
-		exp = 6; /* 10^6 */
-	else if (!strcasecmp(unit, "s") || !strcasecmp(unit, "sec"))
-		exp = 9; /* 10^9 */
-	else if (!strcasecmp(unit, "m") || !strcasecmp(unit, "min"))
-		return parse_min(val, decimal, decimal_places);
-	else
-		pr_warn("The unit '%s' isn't supported\n", unit);
-
-	for (i = 0; i < exp; i++)
-		val *= 10;
-
-	if (decimal) {
-		decimal_places += get_digits(decimal);
-
-		for (i = decimal_places; i < exp; i++)
-			decimal *= 10;
-		val += decimal;
-	}
-	return val;
-}
-
 static bool has_time_unit(const char *str)
 {
 	if (isalpha(str[strlen(str) - 1]))
