+++ utils/utils.h
@@ -14,6 +14,8 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <endian.h>
+#include <string.h>
+#include <ctype.h>
 
 
 #ifndef container_of
@@ -240,5 +242,6 @@ void start_pager(void);
 void wait_for_pager(void);
 
 bool check_time_range(struct uftrace_time_range *range, uint64_t timestamp);
+uint64_t parse_time(char *arg, int limited_digits);
 
 #endif /* __FTRACE_UTILS_H__ */
